requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
psycopg2-binary==2.9.9
Werkzeug==2.0.1
alembic==1.7.1
SQLAlchemy==1.4.23
boto3==1.34.69
Pillow==10.2.0
python-magic==0.4.27
requests==2.31.0 